# openapi.yaml
openapi: 3.0.3
info:
  title: FinMate API
  description: |
    API Contract for FinMate, a Personal Finance & Lifestyle Companion.
    Provides endpoints for Auth, Users, Groups, Expenses, Settlements, Notes, Goals, and Import/Export.

    ### Versioning Strategy
    URL-based versioning is used. All paths start with `/api/v1`.

    ### Pagination & Filtering
    Paginated GET endpoints support standard query parameters:
    * `page`: 1-based page index (default: `1`)
    * `limit`: Page size (default: `20`, max: `100`)
    * `cursor`: Cursor string (for cursor-based pagination on rapidly updated lists like Expenses and Audit Logs)
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local Dev Server
paths:
...
  /users/me:
    get:
      summary: Get current user profile
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile details
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                  - profile
                properties:
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  profile:
                    $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      summary: Update user profile
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 120
                avatarUrl:
                  type: string
                  format: uri
                locale:
                  type: string
                  example: 'en-IN'
                timezone:
                  type: string
                  example: 'Asia/Kolkata'
                defaultCurrency:
                  type: string
                  maxLength: 3
                  example: 'INR'
                monthlyBudget:
                  type: number
                  minimum: 0
                  example: 50000.00
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                  - profile
                properties:
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  profile:
                    $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      summary: Deactivate user account
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deactivated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'VAL_INVALID_INPUT'
            message: 'Input validation failed'
            details:
              - field: 'email'
                issue: 'must be a valid email address'
            retryable: false
    UnauthorizedError:
      description: Authentication failed or token missing/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/users/me'
            errorCode: 'AUTH_TOKEN_EXPIRED'
            message: 'JWT access token has expired'
            retryable: false
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - timestamp
        - path
        - errorCode
        - message
        - retryable
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        errorCode:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - issue
            properties:
              field:
                type: string
              issue:
                type: string
        retryable:
          type: boolean
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, disabled, invited]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Profile:
      type: object
      required:
        - id
        - userId
        - avatarUrl
        - locale
        - timezone
        - defaultCurrency
        - monthlyBudget
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
        timezone:
          type: string
        defaultCurrency:
          type: string
        monthlyBudget:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...